% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo_data.R
\docType{data}
\name{default_logo}
\alias{default_logo}
\title{Logo data for rfetch}
\format{
An object of class \code{character} of length 11.
}
\usage{
default_logo
}
\description{
Logo data for rfetch
}
\keyword{internal}
