% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo_data.R
\name{get_rstudio_logo}
\alias{get_rstudio_logo}
\title{Get logo for RStudio version}
\usage{
get_rstudio_logo(release_name)
}
\arguments{
\item{release_name}{RStudio release name}
}
\value{
Character vector of logo lines
}
\description{
Get logo for RStudio version
}
\keyword{internal}
