% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_theme.R
\docType{package}
\name{rfetch-package}
\alias{rfetch-package}
\title{rfetch: Neofetch‑Style System Information in R}
\description{
Display colourful, ASCII-art blocks (like Neofetch) with key system info inside the R console.
}
\author{
\strong{Maintainer}: Son Tran \email{son.tran@campus.lmu.de}

}
\keyword{internal}
