% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfetch.R
\name{rfetch}
\alias{rfetch}
\title{Print a coloured "fetch" banner}
\usage{
rfetch(
  info = NULL,
  logo = NULL,
  theme = NULL,
  label_col = NULL,
  value_col = NULL,
  logo_col = NULL
)
}
\arguments{
\item{info}{Named list as produced by \code{\link[=rfetch_info]{rfetch_info()}}. If \code{NULL}, the function calls \code{rfetch_info()} internally.}

\item{logo}{Character vector of ASCII lines to prepend.}

\item{theme}{Name of a registered theme for this display only (doesn't change default).}

\item{label_col}{ANSI colour code for labels for this display only.}

\item{value_col}{ANSI colour code for values for this display only.}

\item{fields}{Character vector of field names for this display only.}
}
\value{
Invisibly returns the printed character vector.
}
\description{
Print a coloured "fetch" banner
}
\examples{
# print the "fetch" banne with current default theme
rfetch()

# temporary mono theme
rfetch(theme = "Mono")

# temporary red labels
rfetch(label_col = "31")

# set mono as default
rfetch_theme(name = "Mono")

# now uses mono by default
rfetch()

}
