% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfetch_info.R
\name{rfetch_info}
\alias{rfetch_info}
\title{Collect system information for rfetch}
\usage{
rfetch_info()
}
\value{
A named list with system information (user, host, OS, etc.).
}
\description{
Collect system information for rfetch
}
\examples{
# Get system information
info <- rfetch_info()
str(info)
}
