% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfetch_theme.R
\name{rfetch_theme}
\alias{rfetch_theme}
\title{Register or retrieve rfetch colour themes}
\usage{
rfetch_theme(
  name = NULL,
  label_col = NULL,
  value_col = NULL,
  logo_col = NULL,
  fields = NULL,
  return_pal = FALSE
)
}
\arguments{
\item{name}{Theme name to set as default. Built‑in: "default", "mono", etc.}

\item{label_col}{ANSI colour code for labels (e.g. "36" for cyan) to set as default.}

\item{value_col}{ANSI colour code for values to set as default.}

\item{logo_col}{ANSI colour code for logo to set as default.}

\item{fields}{Character vector of field names to set as default.}

\item{return_pal}{Internal flag: if \code{TRUE}, returns the palette used by \code{\link[=rfetch]{rfetch()}}.}
}
\value{
Invisibly returns a list of registered themes, or (when
\code{return_pal = TRUE}) a palette list.
}
\description{
Register or retrieve rfetch colour themes
}
\examples{
# set Matrix theme as default
rfetch_theme(name = "Matrix")

# create custom theme with labels in red, values in white and logo in green 
rfetch_theme(label_col = "31", value_col = "37", logo_col = "32")
}
